+++ figurelabels/Procfile
@@ -1,1 +1,1 @@
-web: gunicorn app:server
+web: gunicorn index:app.server
