+++ figurelabels/requirements.txt
@@ -4,12 +4,14 @@ chardet==3.0.4
 Click==7.0
 dash==0.42.0
 dash-core-components==0.47.0
+dash-daq==0.1.4
 dash-html-components==0.16.0
 dash-renderer==0.23.0
 dash-table==3.6.0
 decorator==4.4.0
 Flask==1.0.2
 Flask-Compress==1.4.0
+gunicorn==19.9.0
 idna==2.8
 ipython-genutils==0.2.0
 itsdangerous==1.1.0
